Improve modularity of QC modules:
qc_feature_extraction: business logic for extraction of features
qc_features: function extracting qc features
qc_feature_evaluation: business logic for evaluating qc features against
criteria.
